requirements.txt
future
tqdm
matplotlib
numpy
flask
mido
http://download.pytorch.org/whl/cu91/torch-0.3.1-cp35-cp35m-linux_x86_64.whl
tensorflow-gpu
h5py
